fixed variable for calc project, had to take doubles away

+++ project/y.tab.h
@@ -9,7 +9,7 @@
 #ifndef YYSTYPE_IS_DECLARED
 #define YYSTYPE_IS_DECLARED 1
 typedef union {
-  double a;
+  int a;
   char c;
 } YYSTYPE;
 #endif /* !YYSTYPE_IS_DECLARED */
